try moving the idicators
