+++ apps/client/src/features/rundown/event-block/EventBlock.module.scss
@@ -218,6 +218,32 @@ $skip-opacity: 0.1;
   text-align: right;
 }
 
+.indicators {
+  grid-area: next;
+  display: grid;
+  grid-template-areas: 'ind1 ind2 ind3';
+  
+  .indicator {
+    background-color: $black-20;
+    margin: 0 auto;
+    border-radius: 6px;
+    width: 12px;
+    height: 12px;
+  }
+  .indDelay {
+    background-color: $ontime-delay;
+  }
+  .indOverlap {
+    background-color: $violet-600;
+  }
+  .indSpacing {
+    background-color: $blue-600;
+  }
+  .indNextDay {
+    background-color: $gray-600;
+  }
+}
+
 .eventStatus {
   grid-area: status;
   display: flex;
