+++ apps/client/src/features/rundown/event-block/EventBlockInner.tsx
@@ -11,9 +11,12 @@ import { IoPlaySkipForward } from '@react-icons/all-files/io5/IoPlaySkipForward'
 import { IoStop } from '@react-icons/all-files/io5/IoStop';
 import { IoTime } from '@react-icons/all-files/io5/IoTime';
 import { EndAction, Playback, TimerType } from 'ontime-types';
+import { millisToString } from 'ontime-utils';
 
 import TooltipActionBtn from '../../../common/components/buttons/TooltipActionBtn';
 import { useAppMode } from '../../../common/stores/appModeStore';
+import { millisToDelayString } from '../../../common/utils/dateConfig';
+import { cx } from '../../../common/utils/styleUtils';
 import { tooltipDelayMid } from '../../../ontimeConfig';
 import EditableBlockTitle from '../common/EditableBlockTitle';
 import { EventItemActions } from '../RundownEntry';
@@ -103,6 +106,18 @@ const EventBlockInner = (props: EventBlockInnerProps) => {
     playBtnStyles._hover = {};
   }
 
+  const delayedStart = Math.max(0, timeStart + delay);
+  const newTime = millisToString(delayedStart);
+  const delayTime = delay !== 0 ? millisToDelayString(delay) : null;
+
+  const overlap = previousEnd - timeStart;
+  const overlapTime =
+    overlap > 0
+      ? `Overlapping ${millisToDelayString(overlap)}`
+      : overlap < 0
+      ? `Spacing ${millisToDelayString(overlap)}`
+      : null;
+
   return !renderInner ? null : (
     <>
       <EventBlockTimers
@@ -114,10 +129,34 @@ const EventBlockInner = (props: EventBlockInnerProps) => {
         previousEnd={previousEnd}
       />
       <EditableBlockTitle title={title} eventId={eventId} placeholder='Event title' className={style.eventTitle} />
-      {next && (
+      {next ? (
         <Tooltip label='Next event' {...tooltipProps}>
           <span className={style.nextTag}>UP NEXT</span>
         </Tooltip>
+      ) : (
+        <span className={style.indicators}>
+          <Tooltip
+            label={
+              delayTime && (
+                <div>
+                  {delayTime}
+                  <br />
+                  New Time: {newTime}
+                </div>
+              )
+            }
+          >
+            <div className={cx([style.indicator, delayTime ? style.indDelay : ''])} />
+          </Tooltip>
+          <Tooltip label={overlapTime}>
+            <div
+              className={cx([style.indicator, overlap > 0 ? style.indOverlap : overlap < 0 ? style.indSpacing : ''])}
+            />
+          </Tooltip>
+          <Tooltip label='//TODO:'>
+            <div className={cx([style.indicator, timeStart > timeEnd ? style.indNextDay : ''])} />
+          </Tooltip>
+        </span>
       )}
       <EventBlockPlayback
         eventId={eventId}
