+++ apps/client/src/features/rundown/event-block/composite/EventBlockTimers.tsx
@@ -68,16 +68,6 @@ const EventBlockTimers = (props: EventBlockTimerProps) => {
     [timeEnd, timeStart],
   );
 
-  const delayedStart = Math.max(0, timeStart + delay);
-  const newTime = millisToString(delayedStart);
-  const delayTime = delay !== 0 ? millisToDelayString(delay) : null;
-
-  const overlapTime =
-    overlap > 0
-      ? `Overlapping ${millisToDelayString(overlap)}`
-      : overlap < 0
-      ? `Spacing ${millisToDelayString(overlap)}`
-      : null;
 
   return (
     <div className={style.eventTimers}>
@@ -111,39 +101,6 @@ const EventBlockTimers = (props: EventBlockTimerProps) => {
         previousEnd={previousEnd}
         warning={warning.duration}
       />
-      <Tooltip
-        label={
-          <div>
-            {delayTime}
-            <br />
-            New Time: {newTime}
-          </div>
-        }
-      >
-        <div className={cx([style.indicator, delayTime ? style.indDelay : ''])} />
-      </Tooltip>
-      <Tooltip label={overlapTime}>
-        <div className={cx([style.indicator, overlap > 0 ? style.indOverlap : overlap < 0 ? style.indSpacing : ''])} />
-      </Tooltip>
-      <Tooltip label={warning.start}>
-        <div
-          className={cx([style.indicator, warning.start == 'Start time later than end time' ? style.indNextDay : ''])}
-        />
-      </Tooltip>
-      {/* {delayTime && (
-        <div className={style.delayNote}>
-          {delayTime}
-          <br />
-          {`New start: ${newTime}`}
-        </div>
-      )}
-      {overlap != 0 && (
-        <div className={overlap > 0 ? style.spaceNote : style.overlapNote}>
-          {overlap > 0 ? 'Overlap' : 'Space'}
-          <br />
-          {millisToDelayString(overlap)}
-        </div>
-      )} */}
     </div>
   );
 };
